requirements.txt
dynaconf==3.2.11
pandas==2.3.1
pyarrow==21.0.0
pydantic-2.11.7
requests==2.32.4
tqdm==4.67.1
zstandard==0.23.0
